issue_134 Enhancement: mex-functions and matlab functions for ellipsoid approximation and convex hull were written.
git-svn-id: https://ellipsoids.googlecode.com/svn/branches/issue_134_vlazareva@2622 fbfa2205-8925-0410-8e01-63a555b7045b

+++ products/multobj/source/code_files/CH_DAT.c
@@ -44,7 +44,7 @@ void ch_no_file (char *name)
  {printf("%s %s",msg [ch_LNG] [0], name);
  }    /* ch_no_file */
 
-int ch_read_dat (int size,int* indProjVec,int* improveDirectVec,int **objnums)
+int ch_read_dat (int size,double* indProjVec,double* improveDirectVec,int **objnums)
   /* reading the data describing the set */
  {int i, k, nvar, IOstatus;
   char *obj_names, c;
